expression is moved to core
