+++ pyviews/tests/code_tests.py
@@ -4,7 +4,7 @@ from pytest import mark, raises
 
 from pyviews.code import Code, run_code
 from pyviews.core import ViewInfo
-from pyviews.expression import ExpressionError
+from pyviews.core.expression import ExpressionError
 from pyviews.core.rendering import Node
 from pyviews.core.bindable import InheritedDict
 from pyviews.core.xml import XmlNode
